+++ Makefile
@@ -6,7 +6,7 @@
 #    By: gwolf <gwolf@student.42vienna.com>         +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/07/28 13:03:05 by gwolf             #+#    #+#              #
-#    Updated: 2023/08/10 21:40:12 by gwolf            ###   ########.fr        #
+#    Updated: 2023/08/10 22:13:03 by gwolf            ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -108,7 +108,10 @@ SRC :=	utils_memory.c \
 		builtin_export.c \
 		builtin_unset.c \
 		builtin_env.c \
-		builtin_exit.c
+		builtin_exit.c \
+		tkn_list_create.c \
+		tkn_list_delete.c \
+		tkn_list_search.c
 SRCS := $(addprefix $(SRC_DIR)/, $(SRC))
 
 # ******************************
